# backend/openapi-schema.yaml
openapi: 3.0.3
info:
  title: VYBE Music API
  version: v1
  description: Gen-Z Music Discovery Engine REST API
paths:
...
  /api/v1/playlists/{id}/songs/:
    post:
      operationId: playlist_add_song
      description: Add a track to playlist crate. Persists track if needed.
      summary: Add Song to Playlist
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - playlists
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSongToPlaylistRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AddSongToPlaylistRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AddSongToPlaylistRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
          description: ''
  /api/v1/playlists/{id}/songs/{song_id}/:
    delete:
      operationId: playlist_remove_song
      description: Remove a track entry from the playlist.
      summary: Remove Song from Playlist
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: song_id
        schema:
          type: string
        required: true
      tags:
      - playlists
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaylistDetail'
          description: ''
components:
  schemas:
...
    AddSongToPlaylistRequest:
      type: object
      properties:
        youtubeVideoId:
          type: string
          minLength: 1
        youtube_video_id:
          type: string
          minLength: 1
...
    IllustrationEnum:
      enum:
      - chill
      - lockin
      - mainchar
      - unhinged
      - vinyl
      - cassette
      - lofi
      - retro
      type: string
      description: |-
        * `chill` - Chill
        * `lockin` - Lock In
        * `mainchar` - Main Character
        * `unhinged` - Unhinged
        * `vinyl` - Vinyl
        * `cassette` - Cassette
        * `lofi` - Lo-Fi
        * `retro` - Retro
...
    PlaylistDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          maxLength: 255
        description:
          type: string
        author:
          type: object
          additionalProperties: {}
          readOnly: true
        accentColor:
          type: string
        illustration:
          $ref: '#/components/schemas/IllustrationEnum'
        trackCount:
          type: integer
          readOnly: true
        duration:
          type: string
          readOnly: true
        isCurated:
          type: boolean
          readOnly: true
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        tracks:
          type: array
          items: {}
          readOnly: true
      required:
      - accentColor
      - author
      - createdAt
      - duration
      - id
      - isCurated
      - isPublic
      - title
      - trackCount
      - tracks
      - updatedAt
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT